# docs/openapi/slot-database-api.v1.yaml
openapi: 3.0.3
info:
  title: Slot Database API
  version: 1.1.0
  termsOfService: https://slotdatabase.es/developers
  contact:
    name: Slot Database
    url: https://slotdatabase.es/developers
  description: |
    **ES — Slot Database API (Partner Sync v1)**

    Contrato **as-built** de `/api/sync/*` para integradores de cronometraje y
    gestores de pista. Autenticación con API key de usuario (`X-API-Key`).
    Obtén la key en **Perfil** (crear / copiar / regenerar).

    - **Modo personal:** la key actúa sobre el garaje del titular (`/vehicles`,
      `/circuits`, `/timings`).
    - **Modo club:** con key de un admin/owner **o** con la API key de club
      (una por club; el admin la rota en la ficha del club) se listan clubes,
      circuitos, **socios** (`/members`) e invitados (`/guest-members`), y se
      sincronizan competiciones. La key de club no accede al garaje personal.
    - **Idempotencia (opcional):** header `Idempotency-Key` en escrituras
      (sobre todo timings / competiciones). Sin header, el comportamiento
      actual no cambia. Misma key + mismo cuerpo reenvía la primera respuesta.
    - **Rate limit:** por API key, techo alto para una sesión de carrera.
      Si se supera: `429` + `Retry-After`. Los clientes actuales que no
      gestionan 429 no se ven afectados en uso normal.
    - Copia a historial / garaje personal: hace falta **`vehicle_id` asignado**.
      Un `user_id` solo no basta. Pilotos externos sin cuenta son válidos en
...

...

...

...

...

...
servers:
...
security:
  - ApiKeyAuth: []
paths:
  /api/sync/vehicles:
    get:
      tags: [Vehicles]
      operationId: listSyncVehicles
      summary: List garage vehicles
      description: |
        Paginated list of vehicles owned by the API-key user.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/XClientApp'
        - $ref: '#/components/parameters/XClientVersion'
      responses:
        '200':
          description: Vehicle page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleListResponse'
              example:
                vehicles:
                  - id: 11111111-1111-4111-8111-111111111111
                    model: Ferrari 488
                    manufacturer: Scalextric
                    type: GT
                    traction: RWD
                    image: https://example.test/image.jpg
                pagination:
                  total: 1
                  page: 1
                  limit: 25
                  totalPages: 1
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/sync/circuits:
    get:
      tags: [Circuits]
      operationId: listSyncCircuits
      summary: List personal circuits
      parameters:
        - $ref: '#/components/parameters/XClientApp'
        - $ref: '#/components/parameters/XClientVersion'
      responses:
        '200':
          description: Circuits of the API-key user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CircuitListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Circuits]
      operationId: findOrCreateSyncCircuit
      summary: Find or create a personal circuit
      description: |
        Looks up a circuit by name for the API-key user, or creates it.
        Returns the **flat circuit object** (not `{ circuit, created }`).
        `200` if found, `201` if created.
      parameters:
        - $ref: '#/components/parameters/XClientApp'
        - $ref: '#/components/parameters/XClientVersion'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CircuitWrite'
            example:
              name: Pista casa
      responses:
        '200':
          description: Existing circuit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Circuit'
        '201':
          description: Created circuit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Circuit'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/sync/timings:
    get:
      tags: [Timings]
      operationId: listSyncTimings
      summary: List garage timings for a vehicle
      description: |
        `vehicle_id` is required. Optional `circuit_id` / `lane` filter the
        baseline set used by guided training.

        **NOTE:** some clients send `session_type` as a query param; the
        current handler **does not read it**.
      parameters:
        - name: vehicle_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: circuit_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: lane
          in: query
          required: false
          schema:
            type: string
          description: Lane number as string (e.g. `"1"`)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - name: session_type
          in: query
          required: false
          schema:
            type: string
          description: |
            UNKNOWN at server — accepted by some clients, ignored by the handler.
        - $ref: '#/components/parameters/XClientApp'
        - $ref: '#/components/parameters/XClientVersion'
      responses:
        '200':
          description: Filtered timings plus meta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimingListResponse'
        '400':
          description: Missing vehicle_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: vehicle_id es requerido
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Timings]
      operationId: createSyncTiming
      summary: Create a garage timing session
      description: |
        Required: `vehicle_id`, `best_lap_time`, `total_time`, `laps`,
        `average_time`. Times are typically `mm:ss.mmm` (garage is truthy,
        not a strict regex).

        `lap_times` items use `{ lap_number, time_seconds, time_text }`
        (aliases `lap_time_seconds` / `lapNumber` are also accepted).

        Response is the inserted `vehicle_timings` row plus `sync_meta`.
        Extra DB columns may be present.
      parameters:
        - $ref: '#/components/parameters/XClientApp'
        - $ref: '#/components/parameters/XClientVersion'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GarageTimingWrite'
            examples:
              minimal:
                summary: Required fields only
                value:
                  vehicle_id: 11111111-1111-4111-8111-111111111111
                  best_lap_time: '00:12.345'
                  total_time: '02:03.210'
                  laps: 10
                  average_time: '00:12.321'
              withLaps:
                summary: With per-lap times
                value:
                  vehicle_id: 11111111-1111-4111-8111-111111111111
                  best_lap_time: '00:08.210'
                  total_time: '01:22.100'
                  laps: 10
                  average_time: '00:08.210'
                  lane: '1'
                  circuit_id: 22222222-2222-4222-8222-222222222222
                  timing_date: '2026-09-24'
                  session_type: TRAINING
                  best_lap_timestamp: 8.21
                  total_time_timestamp: 82.1
                  average_time_timestamp: 8.21
                  supply_voltage_volts: 12.5
                  reaction_time_ms: 180
                  lap_times:
                    - lap_number: 1
                      time_seconds: 8.3
                      time_text: '00:08.300'
                    - lap_number: 2
                      time_seconds: 8.21
                      time_text: '00:08.210'
      responses:
        '201':
          description: Created timing + sync_meta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GarageTimingCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/sync/clubs/admin:
    get:
      tags: [Clubs]
      operationId: listAdminClubs
      summary: Clubs where the user is admin or owner
      parameters:
        - $ref: '#/components/parameters/XClientApp'
        - $ref: '#/components/parameters/XClientVersion'
      responses:
        '200':
          description: Admin/owner clubs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminClubListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/sync/clubs/{id}/circuits:
    get:
      tags: [Clubs]
      operationId: listClubCircuits
      summary: Club circuits (admin/owner)
      parameters:
        - $ref: '#/components/parameters/ClubId'
        - $ref: '#/components/parameters/XClientApp'
        - $ref: '#/components/parameters/XClientVersion'
      responses:
        '200':
          description: Club circuits
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CircuitListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/sync/clubs/{id}/members:
    get:
      tags: [Clubs]
      operationId: listClubMembers
      summary: Club account members (admin/owner)
      description: |
        **D10.** Roster of socios with a Slot Database account. Admin/owner
        API key required (`403` otherwise). Complements
        `GET /api/sync/clubs/{id}/guest-members` (invitados); this path does
        **not** replace guests.

        Shape mirrors JWT `GET /api/clubs/:id/members` and adds
        `display_name` (public pilot profile, else auth metadata).
      parameters:
        - $ref: '#/components/parameters/ClubId'
        - $ref: '#/components/parameters/XClientApp'
        - $ref: '#/components/parameters/XClientVersion'
      responses:
        '200':
          description: Account members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClubMembersResponse'
              example:
                owner_user_id: 88888888-8888-4888-8888-888888888888
                members:
                  - id: 12121212-1212-4121-8121-121212121212
                    user_id: 88888888-8888-4888-8888-888888888888
                    display_name: Ada Lovelace
                    email: ada@example.test
                    role: admin
                    joined_at: '2026-01-15T10:00:00.000Z'
                    is_owner: true
                  - id: 13131313-1313-4131-8131-131313131313
                    user_id: 99999999-9999-4999-8999-999999999999
                    display_name: Bob
                    email: bob@example.test
                    role: member
                    joined_at: '2026-02-01T10:00:00.000Z'
                    is_owner: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/sync/clubs/{id}/guest-members:
    get:
      tags: [Clubs]
      operationId: listClubGuestMembers
      summary: Club guest members (admin/owner)
      parameters:
        - $ref: '#/components/parameters/ClubId'
        - $ref: '#/components/parameters/XClientApp'
        - $ref: '#/components/parameters/XClientVersion'
      responses:
        '200':
          description: Guest members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestMemberListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/sync/clubs/{id}/guest-members/{guestId}/timings:
    post:
      tags: [Clubs]
      operationId: createGuestMemberTiming
      summary: Record a guest training session
      description: |
        Server requires `circuit_id` + `best_lap_time`. `laps` is optional
        on the server (some clients always send it). Optional vehicle fields
        (`vehicle_id` / `vehicle_model` / `vehicle_type`). Does not require
        `total_time` / `average_time`.
      parameters:
        - $ref: '#/components/parameters/ClubId'
        - name: guestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/XClientApp'
        - $ref: '#/components/parameters/XClientVersion'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuestTimingWrite'
            example:
              circuit_id: 22222222-2222-4222-8222-222222222222
              best_lap_time: '00:11.987'
              laps: 12
              best_lap_timestamp: 11.987
              lane: '2'
              timing_date: '2026-09-24'
              vehicle_id: 11111111-1111-4111-8111-111111111111
      responses:
        '201':
          description: Inserted `club_guest_timings` row
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestTiming'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        User API key from Profile (or `POST /api/auth/api-key`), or a
        club/station key rotated by a club admin. Case-insensitive header
        name. Missing/invalid → `401`. Revoked personal keys are rejected.
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 25
    ClubId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    XClientApp:
      name: X-Client-App
      in: header
      required: false
      schema:
        type: string
      description: |
        Optional client label. When present, may map to `recorded_from`
        on garage inserts. Omitted is valid.
    XClientVersion:
      name: X-Client-Version
      in: header
      required: false
      schema:
        type: string
      description: Optional client version. Not validated by the server.
...
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing:
              value:
                error: No se proporcionó API key. Usa el header X-API-Key.
            invalid:
              value:
                error: API key inválida o expirada
...
    ServerError:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    VehicleListResponse:
      type: object
      required: [vehicles, pagination]
      properties:
        vehicles:
          type: array
          items:
            $ref: '#/components/schemas/Vehicle'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    CircuitListResponse:
      type: object
      required: [circuits]
      properties:
        circuits:
          type: array
          items:
            $ref: '#/components/schemas/Circuit'
...